requirements/base.txt
Django==4.1.7
django-ckeditor==6.5.1
django-phonenumber-field==7.0.2
instaloader==4.9.6
beautifulsoup4==4.11.2
Pillow==9.4.0
pre-commit==3.1.1
requests==2.28.2
django-jazzmin==2.6.0
django-modeltranslation==0.18.9
